requirements.txt
torch~=2.2.2
opencv-python~=4.10.0.82
numpy~=1.24.4
scikit-learn~=1.3.2
tqdm~=4.66.4
torchvision~=0.17.2
pandas~=2.0.3
uvicorn~=0.30.1
fastapi~=0.111.0
pillow~=10.3.0